feat: Allow to delete Rule even not ended yet - MEED-1690 - Meeds-io/MIPs#49 (#1008)
- Prior to this change, for a challenge or automatic rule, the program
owner wasn't able to delete it. This change will allow to delete all
rules independently from its statuses.

- This change will fix creating announcement on deleted Rule and will deny
access to marked as deleted rules from UI.

- Prior to this change, a simple user can directly access a program
through direct URL. This change will ensure to secure program detail
screen access by checking user permissions in REST call.

+++ services/src/main/java/org/exoplatform/addons/gamification/storage/RealizationStorage.java
@@ -182,11 +182,11 @@ public class RealizationStorage {
                                                                         earnerType));
   }
 
-  public Long countRealizationsByRuleId(long ruleId) {
+  public int countRealizationsByRuleId(long ruleId) {
     return gamificationHistoryDAO.countRealizationsByRuleId(ruleId);
   }
 
-  public Long countRealizationsByRuleIdAndEarnerType(long ruleId, IdentityType earnerType) {
+  public int countRealizationsByRuleIdAndEarnerType(long ruleId, IdentityType earnerType) {
     return gamificationHistoryDAO.countRealizationsByRuleIdAndEarnerType(ruleId, earnerType);
   }
